add checkbox result save to Mongo

+++ public/javascripts/wishlist.js
@@ -5,22 +5,27 @@ addWishList?.addEventListener('click', (event) => {
   add?.classList.toggle('invisible');
 });
 
-document.querySelector('#wishlist-main-container')?.addEventListener('click', async (event) => { 
-  event.preventDefault();
-  const wishId = event.target.id;
-  const url = window.location.href.toString().split(window.location.host)[1];
-  const roomId = url.split('/')[2];
-  const response = await fetch(`/rooms/${roomId}/wishlist/${wishId}`, {
-    method: "DELETE"
-  })
-  if (response.status === 200) {
-    const card = event.target.closest('.card');
-    card.remove();
-  } else {}
+// удаление вишлиста
+document
+  .querySelector('#wishlist-main-container')
+  ?.addEventListener('click', async (event) => {
+    if (event.target.classList.contains('delete-weshlist')) {
+      event.preventDefault();
+      const wishId = event.target.id;
+      const url = window.location.href
+        .toString()
+        .split(window.location.host)[1];
+      const roomId = url.split('/')[2];
+      const response = await fetch(`/rooms/${roomId}/wishlist/${wishId}`, {
+        method: 'DELETE',
+      });
+      if (response.status === 200) {
+        const card = event.target.closest('.card');
+        card.remove();
+      }
+    }
+  });
 
-
-
-})
 // отправка на ручку Cоздание нового вишлиста (Категории)
 const sendNewWishlist = document.querySelector('.add-wishlist-button');
 sendNewWishlist?.addEventListener('click', async (event) => {
@@ -50,11 +55,12 @@ wishlist?.addEventListener('click', async (event) => {
     const url = window.location.href.toString().split(window.location.host)[1];
     const roomId = url.split('/')[2];
     const wishlistId = url.split('/')[4].replace('#', '');
-    const name = document.querySelector('.add-input-name').value;
-    const price = document.querySelector('.add-input-price').value;
     const listId = event.target.parentElement.id;
+    const price = document
+      .querySelector(`span[id="${listId}"]`)
+      .innerText.slice(0, -2);
+    const name = event.target.parentElement.firstElementChild.innerText;
     const checkboxValue = event.target.checked;
-    console.log(checkboxValue);
     const response = await fetch(`/rooms/${roomId}/wishlist/${wishlistId}`, {
       method: 'PATCH',
       headers: { 'Content-Type': 'application/json' },
